third_party/blink/web_tests/external/wpt/css/css-grid/masonry/tentative/items/column-automatic-minimum-for-auto-expected.txt

This is a testharness.js-based test.
[FAIL] Check that min-width is honored when sizing auto columns.
  assert_equals: expected "40px" but got "auto"
[FAIL] Check that min-width is honored when sizing auto columns and spanning grid items.
  assert_equals: expected "35px 15px" but got "auto auto"
[FAIL] Check the interactions between width and min-width and auto tracks.
  assert_equals: expected "60px" but got "auto"
[FAIL] Check that borders and paddings are considering when computing min sizes.
  assert_equals: expected "50px" but got "auto"
Harness: the test ran to completion.
